requirements.txt
# Django и основные компоненты
Django==5.1.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.6.0

# База данных
psycopg2-binary==2.9.10

# Валидация и сериализация
django-model-utils==5.0.0
django-extensions==3.2.3

# API документация
drf-spectacular==0.28.0

# Тестирование
pytest==8.3.4
pytest-django==4.9.0
factory-boy==3.3.2

# Качество кода
flake8==7.1.1
black==24.10.0

# Разработка
python-dotenv==1.0.1
django-debug-toolbar==4.4.6

# Работа с датами
python-dateutil==2.9.0 